# Reject query params with an equals sign but no value