Highwinds rejects signed URLs containing a query parameter with an empty value, such as `http://www.example.com/foo?bar=` or `http://www.example.com/foo?bar=&baz=ok`.

However, it is okay to specify a query parameter without an equals sign, such as  `http://www.example.com/foo?bar` or `http://www.example.com/foo?bar&baz=ok`.

This PR introduces a validation step that rejects invalid query parameters.

Also included are some performance optimizations and a benchmark to demonstrate:

## Without string-formatting optimizations
```
goos: darwin
goarch: amd64
pkg: github.com/muxinc/highwinds-urlauth/urlauth
Benchmark_URLAuth_SignURL-12    	 1017592	      1184 ns/op	     680 B/op	      19 allocs/op
PASS
ok  	github.com/muxinc/highwinds-urlauth/urlauth	2.133s
Success: Benchmarks passed.
```

## With string-formatting optimizations
```
goos: darwin
goarch: amd64
pkg: github.com/muxinc/highwinds-urlauth/urlauth
Benchmark_URLAuth_SignURL-12    	 1340202	       919 ns/op	     584 B/op	      14 allocs/op
PASS
ok  	github.com/muxinc/highwinds-urlauth/urlauth	2.095s
Success: Benchmarks passed.
```